websocket: Extract parts of server_connection usable for client
Websocket connection handling is mostly not specific to the server and
thus can be reused on the client easily.
